Reorganize code in backtesting project
