--- ge/analysis/indicators/__init__.py
@@ -1,1 +1,0 @@
-from .__custom_example import MedianIndicator
