+++ trading_strategies/ge/data/__init__.py
@@ -1,0 +1,1 @@
+from .__loader import load_data
