+++ trading_strategies/ge/trading/indicators/__init__.py
@@ -1,0 +1,1 @@
+from .__custom_example import MedianIndicator
