+++ trading_strategies/run_backtest.py
@@ -1,22 +1,23 @@
 # import pandas as pd
-import vectorbt as vbt
 import numpy as np
-from strategies import StrategyBook
-from data_processing import load_data
-import pandas as pd 
+import pandas as pd
+import vectorbt as vbt
+
+from ge.trading.strategy import DMAC
+from ge.data import load_data
+
 
 ohlcv_wbuf = load_data()
 ohlcv_wbuf = ohlcv_wbuf.astype(np.float64)
 
-
 vbt.settings.portfolio['init_cash'] = 100.  # 100$
 vbt.settings.portfolio['fees'] = 0.0025  # 0.25%
 
-# Generamos señales de compra y venta
-entries,exits = StrategyBook.moving_average(ohlcv_wbuf['Close'].values)
+# Generamos señales de compra y venta.
+entries, exits = DMAC(fast_ma=10, slow_ma=20).entries_and_exits(ohlcv_wbuf['Close'].values)
 
-df=pd.concat([ohlcv_wbuf, pd.DataFrame(entries,index=ohlcv_wbuf.index,columns=['entry'])], axis=1)
-df=pd.concat([df, pd.DataFrame(exits,index=ohlcv_wbuf.index,columns=['exit'])], axis=1)
+df = pd.concat([ohlcv_wbuf, pd.DataFrame(entries,index=ohlcv_wbuf.index,columns=['entry'])], axis=1)
+df = pd.concat([df, pd.DataFrame(exits, index=ohlcv_wbuf.index,columns=['exit'])], axis=1)
 
 fig = df['Close'].vbt.plot(trace_kwargs=dict(name='Price'))
 fig = df['entry'].vbt.signals.plot_as_entry_markers(ohlcv_wbuf['Close'], fig=fig)
